Added TMPFS and did some changes for Serial Ports.
Automated the TMPFS installation. #91 
Do some Work to enable access to the Serialports of Synology. Created a startupscript that granted Acess to /dev/ttyACM0 and ttyACM1.  otherwise my ds216 resets the acessrights on reboot. 

To realize these two things, is the Admin Password from the DSM on Installation needed. 

Updated Readme on this two points. 
The needed scripts are in the package.tgz. 

Tested on DS216play with DSM 6.1.7-15284 with openHAB 2.3.003