New generation of packaging files
Basically rewriting all scripts and making everything working, so it works with DSM7.
Everything that makes use of root-rights has been removed and the start/stop mechanism is working the same way as for init.d.

Two helper scripts have been added which provide a backdoor to last two bits that need root-rights:
- Adding world-wide permissions for serial devices to access e.g. Z-Wave adapters
- Adding permissions to create lock files with Java. This is for some reason restricted on DSM7.

Signed-off-by: Thomas Karl Pietrowski <thopiekar@gmail.com>